+++ packages/keto-client-wrapper/test/namespaces.ts
@@ -1,15 +1,29 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import type { Context, Namespace } from '@ory/permission-namespace-types';
+import type {
+  Context,
+  Namespace,
+  SubjectSet,
+  // @ts-expect-error - This is a private type from the internal Ory Keto SDK
+} from '@ory/permission-namespace-types';
 
 class User implements Namespace {}
 
+class Group implements Namespace {
+  related: {
+    members: User[];
+  };
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 class Toy implements Namespace {
   related: {
     owners: User[];
+    puppetmasters: SubjectSet<Group, 'members'>[];
   };
 
   permits = {
-    play: (ctx: Context) => this.related.owners.includes(ctx.subject),
+    play: (ctx: Context) =>
+      this.related.owners.includes(ctx.subject) ||
+      this.related.puppetmasters.includes(ctx.subject),
     break: (ctx: Context) => this.permits.play(ctx),
     steal: (ctx: Context) => !this.permits.play(ctx),
   };