+++ packages/keto-client-wrapper/CHANGELOG.md
@@ -1,3 +1,39 @@
+## 0.2.0 (2024-04-03)
+
+
+### 🚀 Features
+
+- **keto-client-wrapper:** support logic operators ([85bdce3](https://github.com/getlarge/nestjs-ory-integration/commit/85bdce3))
+
+
+### 🩹 Fixes
+
+- **keto-client-wrapper:** do NOT import injected class as type ([18ff575](https://github.com/getlarge/nestjs-ory-integration/commit/18ff575))
+
+
+### ❤️  Thank You
+
+- getlarge @getlarge
+
+## 0.1.0 (2024-02-06)
+
+
+### 🚀 Features
+
+- **keto-client-wrapper:** create unauthorizedFactory option and improve error handling ([cfcd91e](https://github.com/getlarge/nestjs-ory-integration/commit/cfcd91e))
+
+
+### 🩹 Fixes
+
+- export isOryError ([6f4e32f](https://github.com/getlarge/nestjs-ory-integration/commit/6f4e32f))
+
+- instantiate default HTTPException with empty constructor ([a542cbd](https://github.com/getlarge/nestjs-ory-integration/commit/a542cbd))
+
+
+### ❤️  Thank You
+
+- getlarge @getlarge
+
 ## 0.0.2 (2024-01-15)
 
 ### 🩹 Fixes