+++ packages/kratos-client-wrapper/CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.1.1 (2024-04-03)
+
+This was a version bump only for kratos-client-wrapper to align it with other projects, there were no code changes.
+
+## 0.1.0 (2024-02-06)
+
+
+### 🚀 Features
+
+- **kratos-client-wrapper:** create unauthorizedFactory option and improve error handling ([e7e5e53](https://github.com/getlarge/nestjs-ory-integration/commit/e7e5e53))
+
+
+### 🩹 Fixes
+
+- export isOryError ([6f4e32f](https://github.com/getlarge/nestjs-ory-integration/commit/6f4e32f))
+
+- instantiate default HTTPException with empty constructor ([a542cbd](https://github.com/getlarge/nestjs-ory-integration/commit/a542cbd))
+
+
+### ❤️  Thank You
+
+- getlarge @getlarge
+
 ## 0.0.2 (2024-01-15)
 
 This was a version bump only for kratos-client-wrapper to align it with other projects, there were no code changes.